Remote refactor for Writer
