+++ communication/src/main/java/datadog/communication/RemoteFeaturesDiscovery.java
@@ -1,0 +1,8 @@
+package datadog.communication;
+
+public interface RemoteFeaturesDiscovery {
+
+  void discover();
+
+  String getTraceEndpoint();
+}
