+++ communication/src/main/java/datadog/communication/ddagent/DDAgentFeaturesDiscovery.java
@@ -3,6 +3,7 @@ package datadog.communication.ddagent;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
 import com.squareup.moshi.Types;
+import datadog.communication.RemoteFeaturesDiscovery;
 import datadog.communication.http.OkHttpUtils;
 import datadog.communication.monitor.DDAgentStatsDClientManager;
 import datadog.communication.monitor.Monitoring;
@@ -23,7 +24,7 @@ import okhttp3.Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class DDAgentFeaturesDiscovery implements DroppingPolicy {
+public class DDAgentFeaturesDiscovery implements RemoteFeaturesDiscovery, DroppingPolicy {
 
   private static final Logger log = LoggerFactory.getLogger(DDAgentFeaturesDiscovery.class);
