+++ dd-trace-core/src/main/java/datadog/trace/common/writer/PayloadDispatcher.java
@@ -1,5 +1,6 @@
-package datadog.trace.common.writer.ddagent;
+package datadog.trace.common.writer;
 
+import datadog.communication.RemoteFeaturesDiscovery;
 import datadog.communication.ddagent.DDAgentFeaturesDiscovery;
 import datadog.communication.monitor.Monitoring;
 import datadog.communication.monitor.Recording;
@@ -7,7 +8,9 @@ import datadog.communication.serialization.ByteBufferConsumer;
 import datadog.communication.serialization.FlushingBuffer;
 import datadog.communication.serialization.WritableFormatter;
 import datadog.communication.serialization.msgpack.MsgPackWriter;
-import datadog.trace.common.writer.Payload;
+import datadog.trace.common.writer.ddagent.DDAgentApi;
+import datadog.trace.common.writer.ddagent.TraceMapperV0_4;
+import datadog.trace.common.writer.ddagent.TraceMapperV0_5;
 import datadog.trace.core.CoreSpan;
 import datadog.trace.core.monitor.HealthMetrics;
 import java.nio.ByteBuffer;
@@ -21,13 +24,13 @@ public class PayloadDispatcher implements ByteBufferConsumer {
 
   private static final Logger log = LoggerFactory.getLogger(PayloadDispatcher.class);
 
-  private final DDAgentApi api;
-  private final DDAgentFeaturesDiscovery featuresDiscovery;
+  private final RemoteApi api;
+  private final RemoteFeaturesDiscovery featuresDiscovery;
   private final HealthMetrics healthMetrics;
   private final Monitoring monitoring;
 
   private Recording batchTimer;
-  private TraceMapper traceMapper;
+  private RemoteMapper mapper;
   private WritableFormatter packer;
 
   private final FixedSizeStripedLongCounter droppedSpanCount =
@@ -63,34 +66,33 @@ public class PayloadDispatcher implements ByteBufferConsumer {
     // there are alternative approaches to avoid blocking here, such as
     // introducing an unbound queue and another thread to do the IO
     // however, we can't block the application threads from here.
-    if (null == traceMapper || !packer.format(trace, traceMapper)) {
+    if (null == mapper || !packer.format(trace, mapper)) {
       healthMetrics.onFailedPublish(trace.get(0).samplingPriority());
     }
   }
 
   private void selectTraceMapper() {
-    if (null == traceMapper) {
+    if (null == mapper) {
       if (featuresDiscovery.getTraceEndpoint() == null) {
         featuresDiscovery.discover();
       }
       String tracesUrl = featuresDiscovery.getTraceEndpoint();
       if (DDAgentFeaturesDiscovery.V5_ENDPOINT.equalsIgnoreCase(tracesUrl)) {
-        this.traceMapper = new TraceMapperV0_5();
+        this.mapper = new TraceMapperV0_5();
       } else if (null != tracesUrl) {
-        this.traceMapper = new TraceMapperV0_4();
+        this.mapper = new TraceMapperV0_4();
       }
-      if (null != traceMapper && null == packer) {
+      if (null != mapper && null == packer) {
         this.batchTimer =
-            monitoring.newTimer(
-                "tracer.trace.buffer.fill.time", "endpoint:" + traceMapper.endpoint());
-        this.packer = new MsgPackWriter(new FlushingBuffer(traceMapper.messageBufferSize(), this));
+            monitoring.newTimer("tracer.trace.buffer.fill.time", "endpoint:" + mapper.endpoint());
+        this.packer = new MsgPackWriter(new FlushingBuffer(mapper.messageBufferSize(), this));
         batchTimer.start();
       }
     }
   }
 
   Payload newPayload(int messageCount, ByteBuffer buffer) {
-    return traceMapper
+    return mapper
         .newPayload()
         .withBody(messageCount, buffer)
         .withDroppedSpans(droppedSpanCount.getAndReset())
@@ -106,8 +108,8 @@ public class PayloadDispatcher implements ByteBufferConsumer {
       Payload payload = newPayload(messageCount, buffer);
       final int sizeInBytes = payload.sizeInBytes();
       healthMetrics.onSerialize(sizeInBytes);
-      DDAgentApi.Response response = api.sendSerializedTraces(payload);
-      traceMapper.reset();
+      RemoteApi.Response response = api.sendSerializedTraces(payload);
+      mapper.reset();
       if (response.success()) {
         if (log.isDebugEnabled()) {
           log.debug("Successfully sent {} traces to the API", messageCount);
