+++ dd-trace-core/src/main/java/datadog/trace/common/writer/RemoteResponseListener.java
@@ -1,0 +1,8 @@
+package datadog.trace.common.writer;
+
+import java.util.Map;
+
+public interface RemoteResponseListener {
+  /** Invoked after the api receives a response from the remote service. */
+  void onResponse(String endpoint, Map<String, Map<String, Number>> responseJson);
+}
